Enforce consistency naming between tables
Some fields in BAT, RST and SIT have been renamed to match the same name (for
the exact same purpose) in SDT. DVBlast is not impacted but other
applications may have to be changed.

BAT: table ID extension is a bouquet ID, not a network ID.

Correct handling of EIT (service-specific) and BAT (bouquet-specific) in
dvb_print_si. Also free all allocated structures so that we now pass
valgrind memcheck.