Implement support for private data specifier 0x233a for desc 0x83.
I have been looking at implementing support for parsing the LCN
descriptor for UK DVB-T and to my supprise most of it was already there
but for a different standard. In the D-Book LCN is tag 0x83 with private
data specifier of 0x233a where as what is currently implented is the
IEC/CENELEC 62 216 standard where the only difference (in the actual
data) is the latter has 1 bit that specifies if the channel is vissible
or not and the private data specifier is 0x0028.