CI: change image, specify tag (fix pipeline)
The runner's allowed_images pattern does not match an image without
specifying a version. Since administrator access is required in order to
append or modify the runner configuration, this change attempts to
satisfy the existing pattern which expects a ':' followed by a tag.

This also changes the image to run, as it turns out that the
'videolan-base-xenial' image does not have git installed which causes
the pipeline to fail. The image 'vlc-debian-stable' was found to be
suitable, but I have not done an exhaustive search in order to find out
whether it is the optimal choice.

Following this change, the pipeline returns to normal operation which is
to compile every header (i.e the result of `find * -name '*.h'`) in the
repository with both a C and a C++ compiler successfully, or fail.

Co-authored-by: Tristan Matthews <tmatth@videolan.org>